+++ Runtime/MBus/components/MBusUnSubscriberBase.cs
@@ -1,12 +1,14 @@
 using System;
 using UnityEngine;
 
-namespace MBus
+namespace MBus.components
 {
     public class MBusUnSubscriberBase : MonoBehaviour
     {
         internal object Handler;
         internal MBus Bus;
         internal Type Type;
+        internal object Value;
+        internal bool IsValueSubscription;
     }
 }