+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-### 1.1.0-alpha.1
+### 1.1.0-beta.1
 - added value subscriptions: it is possible now to subscribe to a specific value instead of just a type.
 ~~~
 void OnValue() 
@@ -16,8 +16,10 @@ bus.SendMessage(100);
 // won't trigger
 bus.SendMessage(101);
 
-Value subscriptions will not receive a parameter.
+// unsubscripe
+bus.Unsubscribe(OnValue, 100);
 ~~~
+Value subscriptions will not receive a parameter.
 
 ### 1.0.0 
 - Initial release